ausmarton: workaround for persistence of bookmark to tag relations, refactored a bit of controller code deleting tags from a bookmark

+++ CodeIgniter_2.1.3/application/controllers/bookmarks.php
@@ -1,14 +1,13 @@
 <?php if ( ! defined('BASEPATH')) exit('No direct script access allowed');
 
 class Bookmarks extends CI_Controller {
-	
-    	function __construct()
-    	{
-        	parent::__construct();
+
+	function __construct()	{
+		parent::__construct();
 		$this->load->model('Bookmark');
 		$this->load->model('Tag');
-    	}
-//TODO:
+  }
+	//TODO:
 	public function tag($tag_name) {
 		$tag_model = new Tag();
 		$tag_model->where(array('name' => $tag_name));
@@ -16,46 +15,57 @@ class Bookmarks extends CI_Controller {
 		$data['bookmarks'] = $tag->bookmark->get();
 		$this->load->view('bookmarks',$data);
 	}
+//TODO: MERGE WITH TAG CONTROLLER
+	private function tag_to_array($tag) {
+		return array('id' => $tag->id, 'name' => $tag->name);
+	}
 
-	private function get_JSON($bookmark_models) {
-		$bookmarks = array();
-		foreach($bookmark_models as $bookmark_model) {
-			$tags = array();
-			$tag_models = $bookmark_model->tag->get();
-			foreach($tag_models as $tag_model)
-				$tags[] = array(
-					'id' => $tag_model->id,
-					'name' => $tag_model->name
-				);
-			$bookmarks[] = array(
+	private function tags_to_array($tag_models) {
+		$tags = array();
+		foreach($tag_models as $tag_model) {
+			$tags[] = $this->tag_to_array($tag_model);
+		}
+		return $tags;
+	}
+
+	private function bookmark_to_array($bookmark_model) {
+		return array(
 				'id' => $bookmark_model->id,
 				'name' => $bookmark_model->name,
 				'url' => $bookmark_model->url,
-				'tags' => $tags
+				'tags' => $this->tags_to_array($bookmark_model->tag->get()),
 			);
+	}
+
+	private function bookmarks_to_array($bookmark_models) {
+		$bookmarks = array();
+		foreach($bookmark_models as $bookmark_model) {
+			$bookmarks[] = $this->bookmark_to_array($bookmark_model);
 		}
 		return $bookmarks;
 	}
 
-	public function get_index()
-	{
+	public function get_index()	{
 		$bookmark = new Bookmark();
 		$data['bookmarks'] = $bookmark->get();
-		echo json_encode($this->get_JSON($data['bookmarks']));
-		//$this->load->view('bookmarks',$data);
+		echo json_encode($this->bookmarks_to_array($data['bookmarks']));
 	}
+
 	private function get_from_JSON() {
 		return json_decode(file_get_contents("php://input"), true);
 	}
-	public function post_create(){
+
+	public function post_create() {
 		$bookmark = new Bookmark();
 		$this->save($bookmark);
 	}
-	public function put_update($id){
+
+	public function put_update($id) {
 		$bookmark = new Bookmark();
 		$bookmark->id = $id;
 		$this->save($bookmark);
 	}
+
 	private function save($bookmark) {
 		$post_data = $this->get_from_JSON();
 		$bookmark->name = $post_data['name'];
@@ -65,26 +75,29 @@ class Bookmarks extends CI_Controller {
 			$bookmark = $bookmark_model->get_by_id($bookmark->id);
 			$existing_tags = array();
 			foreach($bookmark->tag->get() as $tag)
-				$existing_tags[] = $tag->id;
+				$existing_tags[] = $this->tag_to_array($tag);
 			$new_tags = $post_data['tags'];
-			$tags_to_be_removed = array_diff($existing_tags, $new_tags);				
-			$tags_to_be_added = array_diff($new_tags, $existing_tags);
-			foreach($tags_to_be_removed as $id) {
+			$applicable_tags = $new_tags+$existing_tags;
+			foreach($applicable_tags as $tag_array) {
 				$tag_model = new Tag();
-				$tag = $tag_model->where(array('id' => $id))->get();
-				$tag->delete($bookmark);
+				if(trim($tag_array['id']) == "") {
+					$tag_model->name = $tag_array['name'];
+					$tag_model->save();
+					$tag = $tag_model->where(array('name' => $tag_array['name']))->get();
+				} else
+					$tag = $tag_model->where(array('id' => $tag_array['id']))->get();
+				if(in_array($tag_array,$new_tags)) {
+				  $tag->save($bookmark);
+				} else {
+					$tag->delete($bookmark);
+				}
 			}
-			foreach($tags_to_be_added as $id) {
-				$tag_model = new Tag();
-				$tag = $tag_model->where(array('id' => $id))->get();
-				$tag->save($bookmark);
-			}
-			//redirect(site_url() . "/bookmarks/",'location');
+			$bookmark = $bookmark->where(array('id' => $bookmark->id))->get();
+			echo json_encode($this->bookmark_to_array($bookmark));
 		}
 	}
 
-	public function get_new()
-	{
+	public function get_new()	{
 		$data['type'] = "create";
 		$data['name'] = "";
 		$data['url'] = "";
@@ -92,6 +105,7 @@ class Bookmarks extends CI_Controller {
 		$data['tags'] = array();
 		$this->load->view('save_bookmark',$data);
 	}
+
 	private function all_available_tags_for_dropdown() {
 		$all_tags = array();
 		$tag_model = new Tag();
@@ -100,33 +114,28 @@ class Bookmarks extends CI_Controller {
 		return $all_tags;	
 	}
 
-	public function get_edit($id)
-	{
+	public function get_edit($id)	{
 		$data['type'] = "edit";
 		$this->load->model('bookmark');
 		$bookmark_model = new Bookmark();
 		$bookmark = $bookmark_model->get_by_id($id);
-		
+
 		$tags = array();
 		foreach($bookmark->tag->get() as $tag)
 			$tags[] = $tag->id;
 		$data['tags'] = $tags;
 		$data['all_tags'] = $this->all_available_tags_for_dropdown();
-		
+
 		$data['id'] = $bookmark->id;
 		$data['name'] = $bookmark->name;
 		$data['url'] = $bookmark->url;
 		$this->load->view('save_bookmark',$data);
 	}
 
-	public function delete($id)
-	{
+	public function delete($id)	{
 		$bookmark = new Bookmark();
 		$bookmark->id = $id;
 		$bookmark->delete();
-			//redirect(site_url() . "/bookmarks/",'location');
+		//redirect(site_url() . "/bookmarks/",'location');
 	}
 }
-
-/* End of file welcome.php */
-/* Location: ./application/controllers/welcome.php */
